Parse the remote sha _after_ pulling
Previously, we were parsing the remote sha out of the upstream branch
before pulling. This meant that when we later went to print a log of the
changes since that sha, it resulted in a no-op because our remote hadn't
moved on yet.

The simplest fix for this is to parse the remote sha after we perform
our pull, and right before we want to print the log.